feat: Provide a way to initialize RendererIpc with nodeIntegration disabled, contextIsolation enabled
Unlike v1 implementation, the class instance is not passeable through Context Bridge at all. The problem in V1 was that event listeners could not be removed when it's passed via context bridge. The new setup allows initializing RendererIpc in the main execution context, with ipcRenderer passed through context bridge. This allows removing event listeners (it did not work in V1), however, the ipc instance cannot be fully destroyed (which is an acceptable caveat given the temporary nature of a renderer).